requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
click==8.1.8
discord.py==2.5.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
